pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automata-viewer"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib==3.8.3",
    "PyQt6==6.9.1",
    "PyYAML==6.0.2",
    "userpaths==0.1.3"
]

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
automata_builder = "automata_builder.__main__:main"

[tool.setuptools.package-data]
automata_builder = [
    "styles/*",
    "locale/*",
]
